improved test suite
