+++ test/binoptest.h
@@ -1,0 +1,3 @@
+#pragma once
+
+int test_binop(void);
