+++ test/compilertest.c
@@ -1,480 +1,19 @@
-#include "compiler/ast/ast.h"
-#include "compiler/parser/parser.h"
-#include <compiler/compiler/compiler.h>
-#include <color.h>
+#include "opcode.h"
 #include "yats.h"
 #include "compilertest.h"
+#include "color.h"
+#include "binoptest.h"
+#include "unoptest.h"
+#include "literaltest.h"
 
-void setup_compiler(char *file_contents) {
-    Parser *parser = parser_new(setup_lexer(file_contents));
-    Compiler *compiler = compiler_new(parser, "dump.yb");
-    compile(compiler);
-    compiler_del(compiler);
-}
+#define RUN(test) __YASL_TESTS_FAILED__ |= test()
 
 SETUP_YATS();
-
-#define ASSERT_BC_EQ(left, right) do {\
-    if (sizeof(left) == sizeof(right) && !memcmp(left, right, sizeof(left))) {\
-        /*printf(K_GRN "assert passed in %s: line %d" K_END "\n", __func__, __LINE__);*/\
-    } else {\
-        printf(K_RED "assert failed in %s: line %d" K_END "\n", __func__, __LINE__);\
-        puts("expected: ");\
-        for (int i = 0; i < sizeof(left); i++) printf(i < sizeof(right) && (left)[i] == (right)[i] ? K_GRN "%02x " : K_RED "%02x ", (left)[i] & 0xFF);\
-        printf(K_END "\n");\
-        puts("actual: ");\
-        for (int i = 0; i < sizeof(right); i++) printf(i < sizeof(left) && (left)[i] == (right)[i] ? K_GRN "%02x " : K_RED "%02x ", (right)[i] & 0xFF);\
-        printf(K_END "\n");\
-        __YASL_TESTS_FAILED__ = 1;\
-    }\
-} while(0)
-
-#define ASSERT_GEN_BC_EQ(expected, fc) do{\
-    remove("dump.yb");\
-    setup_compiler(fc);\
-    FILE *file = fopen("dump.yb", "rb");\
-    int64_t size = getsize(file);\
-    unsigned char actual[size];\
-    fread(actual, sizeof(char), size, file);\
-    ASSERT_BC_EQ(expected, actual);\
-    fclose(file);\
-} while(0)
-
-static int64_t getsize(FILE *file) {
-    fseek(file, 0, SEEK_END);
-    int64_t size = ftell(file);
-    fseek(file, 0, SEEK_SET);
-    return size;
-}
 
 // NOTE: these tests depend on the endianess of the system, so they may fail on big endian systems.
 
 /// Literals
 ////////////////////////////////////////////////////////////////////////////////
-
-static void test_undef() {
-    unsigned char expected[] = {
-            0x10, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            NCONST,
-            POP,
-            HALT
-    };
-    ASSERT_GEN_BC_EQ(expected, "undef;");
-}
-
-static void test_true() {
-    unsigned char expected[] = {
-            0x10, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            BCONST_T,
-            POP,
-            HALT
-    };
-    ASSERT_GEN_BC_EQ(expected, "true;");
-}
-
-static void test_false() {
-    unsigned char expected[] = {
-            0x10, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            BCONST_F,
-            POP,
-            HALT
-    };
-    ASSERT_GEN_BC_EQ(expected, "false;");
-}
-
-static void test_bin() {
-    unsigned char expected[] = {
-            0x10, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            ICONST,
-            0x03, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            POP,
-            HALT
-    };
-    ASSERT_GEN_BC_EQ(expected, "0b11;");
-}
-
-static void test_oct() {
-    unsigned char expected[] = {
-            0x10, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            ICONST,
-            0x0A, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            POP,
-            HALT
-    };
-    ASSERT_GEN_BC_EQ(expected, "0o12;");
-}
-
-static void test_dec() {
-    unsigned char expected[]  = {
-            0x10, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            ICONST,
-            0x0A, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            POP,
-            HALT
-    };
-    ASSERT_GEN_BC_EQ(expected, "10;");
-}
-
-static void test_hex() {
-    unsigned char expected[] = {
-            0x10, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            ICONST,
-            0x10, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            POP,
-            HALT
-    };
-    ASSERT_GEN_BC_EQ(expected, "0x10;");
-}
-
-static void test_float() {
-    unsigned char expected[] = {
-            0x10, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            DCONST,
-            0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xF0, 0x3F,
-            POP,
-            HALT
-    };
-    ASSERT_GEN_BC_EQ(expected, "1.0;");
-}
-
-static void test_string() {
-    unsigned char expected[] = {
-            0x1C, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            0x04, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            'Y', 'A', 'S', 'L',
-            NEWSTR,
-            0x10, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            POP,
-            HALT
-    };
-    ASSERT_GEN_BC_EQ(expected, "'YASL';");
-}
-
-static void test_list() {
-    unsigned char expected[] = {
-            0x10, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            END,
-            ICONST,
-            0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            ICONST,
-            0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            ICONST,
-            0x02, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            ICONST,
-            0x03, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            ICONST,
-            0x04, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            NEWLIST,
-            POP,
-            HALT
-    };
-    ASSERT_GEN_BC_EQ(expected, "[0, 1, 2, 3, 4];");
-}
-
-static void test_table() {
-    unsigned char expected[] = {
-            0x27, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            0x04, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            'z',  'e',  'r',  'o',
-            0x03, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            'o',  'n',  'e',
-            END,
-            ICONST,
-            0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            NEWSTR,
-            0x10, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            ICONST,
-            0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            NEWSTR,
-            0x1C, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            NEWTABLE,
-            POP,
-            HALT
-    };
-    ASSERT_GEN_BC_EQ(expected, "[0:'zero', 1:'one'];");
-}
-
-/// Unary Operators
-////////////////////////////////////////////////////////////////////////////////
-
-static void test_neg() {
-    unsigned char expected[] = {
-            0x10, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            ICONST,
-            0x10, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            NEG,
-            POP,
-            HALT
-    };
-    ASSERT_GEN_BC_EQ(expected, "-16;");
-}
-
-static void test_len() {
-    unsigned char expected[] = {
-            0x1C, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            0x04, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            'Y', 'A', 'S', 'L',
-            NEWSTR,
-            0x10, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            LEN,
-            POP,
-            HALT
-    };
-    ASSERT_GEN_BC_EQ(expected, "#'YASL';");
-}
-
-static void test_not() {
-    unsigned char expected[] = {
-            0x10, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            BCONST_T,
-            NOT,
-            POP,
-            HALT
-    };
-    ASSERT_GEN_BC_EQ(expected, "!true;");
-}
-
-static void test_bnot() {
-    unsigned char expected[] = {
-            0x10, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            ICONST,
-            0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            BNOT,
-            POP,
-            HALT
-    };
-    ASSERT_GEN_BC_EQ(expected, "^0x00;");
-}
-
-/// Binary Operators
-////////////////////////////////////////////////////////////////////////////////
-
-static void test_mul() {
-    unsigned char expected[] = {
-            0x10, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            ICONST,
-            0x02, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            ICONST,
-            0x03, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            MUL,
-            POP,
-            HALT
-    };
-    ASSERT_GEN_BC_EQ(expected, "2 * 3;");
-}
-
-static void test_fdiv() {
-    unsigned char expected[] = {
-            0x10, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            ICONST,
-            0x03, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            ICONST,
-            0x02, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            FDIV,
-            POP,
-            HALT
-    };
-    ASSERT_GEN_BC_EQ(expected, "3 / 2;");
-}
-
-static void test_idiv() {
-    unsigned char expected[] = {
-            0x10, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            ICONST,
-            0x03, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            ICONST,
-            0x02, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            IDIV,
-            POP,
-            HALT
-    };
-    ASSERT_GEN_BC_EQ(expected, "3 // 2;");
-}
-
-static void test_mod() {
-    unsigned char expected[] = {
-            0x10, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            ICONST,
-            0x05, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            ICONST,
-            0x03, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            MOD,
-            POP,
-            HALT
-    };
-    ASSERT_GEN_BC_EQ(expected, "5 % 3;");
-}
-
-static void test_add() {
-    unsigned char expected[] = {
-            0x10, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            ICONST,
-            0x05, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            ICONST,
-            0x05, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            ADD,
-            POP,
-            HALT
-    };
-    ASSERT_GEN_BC_EQ(expected, "5 + 5;");
-}
-
-static void test_sub() {
-    unsigned char expected[] = {
-            0x10, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            ICONST,
-            0x05, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            ICONST,
-            0x03, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            SUB,
-            POP,
-            HALT
-    };
-    ASSERT_GEN_BC_EQ(expected, "5 - 3;");
-}
-
-static void test_bshl() {
-    unsigned char expected[] = {
-            0x10, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            ICONST,
-            0x02, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            ICONST,
-            0x03, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            BSL,
-            POP,
-            HALT
-    };
-    ASSERT_GEN_BC_EQ(expected, "2 << 3;");
-}
-
-static void test_bshr() {
-    unsigned char expected[] = {
-            0x10, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            ICONST,
-            0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            ICONST,
-            0x02, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            BSR,
-            POP,
-            HALT
-    };
-    ASSERT_GEN_BC_EQ(expected,"8 >> 2;");
-}
-
-static void test_band() {
-    unsigned char expected[] = {
-            0x10, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            ICONST,
-            0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            ICONST,
-            0x02, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            BAND,
-            POP,
-            HALT
-    };
-    ASSERT_GEN_BC_EQ(expected,"8 & 2;");
-}
-
-static void test_bxor() {
-    unsigned char expected[] = {
-            0x10, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            ICONST,
-            0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            ICONST,
-            0x02, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            BXOR,
-            POP,
-            HALT
-    };
-    ASSERT_GEN_BC_EQ(expected,"8 ^ 2;");
-}
-
-static void test_bor() {
-    unsigned char expected[] = {
-            0x10, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            ICONST,
-            0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            ICONST,
-            0x02, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            BOR,
-            POP,
-            HALT
-    };
-    ASSERT_GEN_BC_EQ(expected,"8 | 2;");
-}
-
-static void test_concat() {
-    unsigned char expected[] = {
-            0x10, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            ICONST,
-            0x02, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            ICONST,
-            0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            CNCT,
-            POP,
-            HALT
-    };
-    ASSERT_GEN_BC_EQ(expected,"2 ~ 1;");
-}
-
-static void test_and() {
-    unsigned char expected[] = {
-            0x10, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            BCONST_T,
-            DUP,
-            BRF_8,
-            0x02, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            POP,
-            BCONST_F,
-            POP,
-            HALT
-    };
-    ASSERT_GEN_BC_EQ(expected,"true && false;");
-}
-
-static void test_or() {
-    unsigned char expected[] = {
-            0x10, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            BCONST_T,
-            DUP,
-            BRT_8,
-            0x02, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            POP,
-            BCONST_F,
-            POP,
-            HALT
-    };
-    ASSERT_GEN_BC_EQ(expected,"true || false;");
-}
 
 /// Control Flow
 ////////////////////////////////////////////////////////////////////////////////
@@ -557,38 +96,9 @@ static void test_while() {
 ////////////////////////////////////////////////////////////////////////////////
 
 int compilertest() {
-    // Literals
-    test_undef();
-    test_true();
-    test_false();
-    test_bin();
-    test_oct();
-    test_dec();
-    test_hex();
-    test_float();
-    test_string();
-    test_list();
-    test_table();
-    // Unary Operators
-    test_neg();
-    test_len();
-    test_not();
-    test_bnot();
-    // Binary Operators
-    test_mul();
-    test_fdiv();
-    test_idiv();
-    test_mod();
-    test_add();
-    test_sub();
-    test_bshl();
-    test_bshr();
-    test_band();
-    test_bxor();
-    test_bor();
-    test_concat();
-    test_and();
-    test_or();
+    RUN(test_literal);
+    RUN(test_unop);
+    RUN(test_binop);
     // Control Flow
     test_if();
     test_ifelse();
