+++ test/literaltest.h
@@ -1,0 +1,3 @@
+#pragma once
+
+int test_literal(void);
